Extract repeat slot simulation module
